# frontend/openapi/backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "energy-agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/diagnosis/sessions": {
      "get": {
        "tags": [
          "diagnosis-queries"
        ],
        "summary": "Sessions",
        "operationId": "sessions_api_v1_diagnosis_sessions_get",
        "parameters": [
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phase"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "name": "alarm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alarm Id"
            }
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created By"
            }
          },
          {
            "name": "latest_review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latest Review Status"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Time"
            }
          },
          {
            "name": "to_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Time"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisSessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_v1_diagnosis_sessions_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/sessions/{session_id}/timeline": {
      "get": {
        "tags": [
          "diagnosis-queries"
        ],
        "summary": "Timeline",
        "operationId": "timeline_api_v1_diagnosis_sessions__session_id__timeline_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/diagnosis/chat": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_diagnosis_chat_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Session Message",
        "operationId": "session_message_api_v1_diagnosis_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/sessions/{session_id}/messages/stream": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Stream Message",
        "operationId": "stream_message_api_v1_diagnosis_sessions__session_id__messages_stream_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/sessions/{session_id}": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_v1_diagnosis_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SessionSource"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "alarm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Id"
          },
          "alarm_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "CreateSessionRequest"
      },
      "CreateSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "phase": {
            "$ref": "#/components/schemas/DiagnosisPhase"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "merged": {
            "type": "boolean",
            "title": "Merged",
            "default": false
          },
          "duplicate_count": {
            "type": "integer",
            "title": "Duplicate Count",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "run_id",
          "phase",
          "trace_id"
        ],
        "title": "CreateSessionResponse"
      },
...
      "DiagnosisChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "clarification_answers": {
            "items": {
              "$ref": "#/components/schemas/ClarificationAnswer"
            },
            "type": "array",
            "title": "Clarification Answers"
          },
          "expected_memory_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Memory Revision"
          },
          "followup_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "DiagnosisChatRequest"
      },
...
      "DiagnosisPhase": {
        "type": "string",
        "enum": [
          "INIT",
          "PLAN_READY",
          "DATA_FETCHING",
          "EVIDENCE_READY",
          "NEED_USER_INPUT",
          "DRAFT_READY",
          "REVIEWING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "DiagnosisPhase"
      },
      "DiagnosisResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "phase": {
            "$ref": "#/components/schemas/DiagnosisPhase"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagnosisIntent"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredDiagnosisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "tool_summaries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Summaries"
          },
          "clarification_questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            },
            "type": "array",
            "title": "Clarification Questions"
          },
          "degraded_components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded Components"
          },
          "memory_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Revision"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "case_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "run_id",
          "trace_id",
          "phase"
        ],
        "title": "DiagnosisResponse"
      },
...
      "DiagnosisSessionItem": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "alarm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Id"
          },
          "alarm_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Name"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "latest_review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Review Status"
          },
          "final_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Summary"
          },
          "diagnosis_template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Template Id"
          },
          "diagnosis_template_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Template Version"
          },
          "alarm_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Category"
          },
          "guardrail_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardrail Status"
          },
          "failure_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Category"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "run_id",
          "source",
          "phase",
          "risk_level",
          "trace_id",
          "created_at",
          "updated_at"
        ],
        "title": "DiagnosisSessionItem"
      },
      "DiagnosisSessionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DiagnosisSessionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DiagnosisSessionListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "clarification_answers": {
            "items": {
              "$ref": "#/components/schemas/ClarificationAnswer"
            },
            "type": "array",
            "title": "Clarification Answers"
          },
          "expected_memory_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Memory Revision"
          },
          "followup_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SessionMessageRequest"
      },
      "SessionSource": {
        "type": "string",
        "enum": [
          "alarm",
          "chat"
        ],
        "title": "SessionSource"
      },
...
      "TimelineResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "history_complete": {
            "type": "boolean",
            "title": "History Complete"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "history_complete",
          "items"
        ],
        "title": "TimelineResponse"
      },
...
    }
  }
}